refactor: avoid utils
